# shared/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Team4s v3 API
  version: 0.1.0
  description: |
    Baseline OpenAPI contract for current endpoints, including admin content-management routes.
  license:
    name: Proprietary
    url: https://team4s.dev/license
servers:
  - url: /
    description: Relative API base
...
security: []
paths:
...
  /api/v1/auth/issue:
    post:
      tags: [Auth]
      summary: Issue access and refresh token pair
      description: |
        Trusted identity source contract:
        - Non-dev deployments: `Authorization: Bearer <signed token>` is required.
        - Local development only: server-side fallback identity may be used when `AUTH_ISSUE_DEV_MODE=true`.
        - `X-Auth-Issue-Key` is only evaluated for the dev fallback path when `AUTH_ISSUE_DEV_KEY` is configured.
      operationId: issueAuthToken
      parameters:
        - in: header
          name: Authorization
          required: false
          schema:
            type: string
          description: |
            Trusted identity source for non-dev deployments (`Bearer <signed token>` required outside local dev fallback mode).
        - in: header
          name: X-Auth-Issue-Key
          required: false
          schema:
            type: string
          description: Optional local dev fallback gate key; ignored unless `AUTH_ISSUE_DEV_MODE=true` and `AUTH_ISSUE_DEV_KEY` is configured.
      responses:
        "201":
          description: Issued access and refresh token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          description: Missing trusted identity source for active profile or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                required:
                  value:
                    error:
                      message: anmeldung erforderlich
                invalidToken:
                  value:
                    error:
                      message: ungueltiges zugriffstoken
        "503":
          description: Auth state unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authStateUnavailable:
                  value:
                    error:
                      message: auth-status voruebergehend nicht verfuegbar
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access and refresh token pair
      operationId: refreshAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
            examples:
              default:
                value:
                  refresh_token: <refresh-token>
      responses:
        "200":
          description: Rotated access and refresh token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidBody:
                  value:
                    error:
                      message: ungueltiger request body
                missingRefreshToken:
                  value:
                    error:
                      message: refresh_token ist erforderlich
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRefreshToken:
                  value:
                    error:
                      message: ungueltiges refresh-token
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/revoke:
    post:
      tags: [Auth]
      summary: Revoke active access token and session
      operationId: revokeAuthToken
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRevokeRequest"
            examples:
              default:
                value:
                  refresh_token: <refresh-token>
      responses:
        "204":
          description: Session revoked
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidBody:
                  value:
                    error:
                      message: ungueltiger request body
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authRequired:
                  value:
                    error:
                      message: anmeldung erforderlich
                invalidToken:
                  value:
                    error:
                      message: ungueltiges zugriffstoken
        "503":
          description: Auth state unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authStateUnavailable:
                  value:
                    error:
                      message: auth-status voruebergehend nicht verfuegbar
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/anime:
    get:
      tags: [Anime]
      summary: List anime
      operationId: listAnime
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 24
        - name: q
          in: query
          description: Search string (substring match) applied to title fields.
          schema:
            type: string
            maxLength: 100
            example: attack
        - name: letter
          in: query
          description: Uppercase A-Z, "0" for numeric titles, or omitted for all.
          schema:
            type: string
            pattern: '^[A-Z0]|$'
            example: A
        - name: content_type
          in: query
          schema:
            $ref: "#/components/schemas/ContentType"
        - name: status
          in: query
          description: Optional status filter. "disabled" values are only returned when include_disabled=true.
          schema:
            $ref: "#/components/schemas/AnimeStatus"
        - name: fansub_id
          in: query
          description: Optional filter for anime that are linked to a specific fansub group.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: include_disabled
          in: query
          description: Include disabled/preseed anime rows in responses. Default false.
          schema:
            type: boolean
            default: false
        - name: has_cover
          in: query
          description: Optional filter for cover_image presence. true = only records with cover_image, false = only records without cover_image.
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated anime list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnimeResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidPage:
                  value:
                    error:
                      message: ungueltiger page parameter
                invalidQuery:
                  value:
                    error:
                      message: ungueltiger q parameter
                invalidFansubID:
                  value:
                    error:
                      message: ungueltiger fansub_id parameter
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/anime/{id}:
    get:
      tags: [Anime]
      summary: Get anime detail
      operationId: getAnimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: include_disabled
          in: query
          description: Include disabled/preseed anime rows in detail lookup. Default false.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Anime detail payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeDetailResponse"
        "400":
          description: Invalid anime id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  value:
                    error:
                      message: ungueltige anime id
        "404":
          description: Anime not found or disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error:
                      message: anime nicht gefunden
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/anime/{id}/backdrops:
    get:
      tags: [Anime]
      summary: Resolve Jellyfin backdrop and theme-video candidates for anime detail background
      operationId: listAnimeBackdrops
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Backdrop and theme-video candidate URLs (can be empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeBackdropResponse"
        "400":
          description: Invalid anime id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  value:
                    error:
                      message: ungueltige anime id
        "404":
          description: Anime not found or disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error:
                      message: anime nicht gefunden
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/anime/{id}/comments:
    get:
      tags: [Comments]
      summary: List comments for anime
      operationId: listAnimeComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated comment list for anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCommentResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  value:
                    error:
                      message: ungueltige anime id
                invalidPage:
                  value:
                    error:
                      message: ungueltiger page parameter
                invalidPerPage:
                  value:
                    error:
                      message: ungueltiger per_page parameter
        "404":
          description: Anime not found or disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error:
                      message: anime nicht gefunden
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Comments]
      summary: Create comment for anime
      operationId: createAnimeComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateRequest"
            examples:
              default:
                value:
                  content: Danke fuer den Release!
      responses:
        "201":
          description: Created comment payload
          headers:
            X-Comment-RateLimit-Degraded:
              description: Present with `true` when rate limit store is unavailable and limiter runs in degraded mode.
              schema:
                type: string
                enum: ["true"]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentCreateResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  value:
                    error:
                      message: ungueltige anime id
                invalidBody:
                  value:
                    error:
                      message: ungueltiger request body
                missingContent:
                  value:
                    error:
                      message: content ist erforderlich
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authRequired:
                  value:
                    error:
                      message: anmeldung erforderlich
                invalidToken:
                  value:
                    error:
                      message: ungueltiges zugriffstoken
        "404":
          description: Anime not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error:
                      message: anime nicht gefunden
        "503":
          description: Auth state unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                authStateUnavailable:
                  value:
                    error:
                      message: auth-status voruebergehend nicht verfuegbar
        "429":
          description: Rate limit exceeded for comment creation
          headers:
            Retry-After:
              description: Seconds until the next comment create request is allowed.
              schema:
                type: integer
                format: int32
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                rateLimitExceeded:
                  value:
                    error:
                      message: zu viele anfragen, bitte spaeter erneut versuchen
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/fansub-slugs/{slug}:
    get:
      tags: [Fansubs]
      summary: Get fansub group by slug
      operationId: getFansubGroupBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        "200":
          description: Fansub group payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FansubGroupResponse"
        "400":
          description: Invalid fansub slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Fansub group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SignedToken
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
    AuthRefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthRevokeRequest:
      type: object
      properties:
        refresh_token:
          type: string
...
    AuthTokenResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthTokenData"
...
    AnimeStatus:
      type: string
      enum: [disabled, ongoing, done, aborted, licensed]
    ContentType:
      type: string
      enum: [anime, hentai]
...
    PaginatedAnimeResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AnimeListItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    AnimeDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AnimeDetail"
...
    AnimeBackdropResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AnimeBackdropManifest"
...
    CommentCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
    CommentCreateResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CommentListItem"
    PaginatedCommentResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CommentListItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    FansubGroupResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/FansubGroup"
...